Map related function moved to different file
 - these function were moved to tilemgr.cpp:
   map_invalidate_selection_rect
   map_invalidate_map_selection_tiles
