+++ src/openloco/map/tilemgr.h
@@ -9,4 +9,6 @@ namespace openloco::map::tilemgr
     tile get(map_pos pos);
     tile get(coord_t x, coord_t y);
     std::tuple<int16_t, int16_t> get_height(coord_t x, coord_t y);
+    void map_invalidate_map_selection_tiles();
+    void map_invalidate_selection_rect();
 }
